Instantiate Issuer
Instantiates Issuer based of data provided by GET-request sent to
OPENID_CONFIGURATION_URL environment variable. Adds dotenv-dependency
to the project for getting environment variables when running in development.

Made a separate server-directory for existing logic, and ensure we don't
start server before Issuer is instantiated, which is also being provided
to the server.